Implemented katex with a mixin and component
Any component that will render latex should use the new mixin
parse-for-latex-component-mixin. This mixin adds a lifecycle hook to the
component to make sure that the LaTeX gets rendered, even with async
calls.

I used this mixin in markdown-file-latex component -- this extends the
markdown-file component with the mixin above. Works!